Render PKBOARDING type as an apple PKPASS
The spec for rendering the files is here:

https://developer.apple.com/library/archive/documentation/UserExperience/Conceptual/PassKit_PG/Creating.html

If the type is PKBOARDING, render as a PKPASS using the hints to identify the regions the tags should go in.

The main difference between pkpass rendering and espass rendering is
that the espass fields are all rendered linearly, however the pkpass
fields need to be assigned to a region (using the new espass hint
parameter) and then rendered vertically.  By convention, the
primaryFields only contains the to and from airports, so some special
magic is applied to them to get them to render larger and in a left
and right justified pair.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>