Fix temp file:java.ioIOExcepction: Stream Closed when loading a pkpass (issue #431)
This seems to be occurring because the InputStreamProvider tries to
wrap the InputStream in a .use when constructing a fromURI from a
content:// URI.  This seems no-longer to work on androids 10 and 11 at
least.  Fix this by simply wrappering an ordinary inputStream (as is
done for file:// and https://) and instead doing the correct .use
directive at the point of use (so the content:// URI is correctly
closed).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>